refactor: 예외 핸들러 코드 이동
